add framework net60 and cached the all typeof()
